+++ CHANGELOG.rst
@@ -2,6 +2,10 @@
  Usagi Changelog
 =================
 
+Changes since version 0.3.0
+===========================
+
+
 Version 0.3.0
 =============